pyproject.toml
[project]
name = "triple-kepler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "yfinance>=0.2.66",
]
